refactored fade in section
