+++ Reviews/src/Components/ReviewList/ReviewList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ReviewListItem from './ReviewListItem.jsx';
-import FadeInSection from '../FadeInTest.jsx';
+import FadeInSection from '../FadeInSection.jsx';
 
 const ReviewList = ({
   reviews, reviewCount, markAsHelpful, reportReview, photoModal, sendClickData,
@@ -9,7 +9,8 @@ const ReviewList = ({
   const useTheseReviews = reviews.slice(0, reviewCount);
   const old = useTheseReviews.slice(0, useTheseReviews.length - 2);
   const newRevs = useTheseReviews.slice(useTheseReviews.length - 2);
-  const setUpReviews = (arr) => arr.map((review) => (
+
+  const revListItem = (review) => (
     <ReviewListItem
       markAsHelpful={markAsHelpful}
       review={review}
@@ -18,15 +19,25 @@ const ReviewList = ({
       photoModal={photoModal}
       sendClickData={sendClickData}
     />
-  ));
+  );
+
+  const setUpReviews = (arr) => arr.map((review) => {
+    if (arr === newRevs) {
+      return (
+        <FadeInSection>
+          {revListItem(review)}
+        </FadeInSection>
+      );
+    }
+    return (
+      revListItem(review)
+    );
+  });
+
   return (
     <div id="reviewList">
-
       {setUpReviews(old)}
-      <FadeInSection>
-        {setUpReviews(newRevs)}
-
-      </FadeInSection>
+      {setUpReviews(newRevs)}
     </div>
   );
 };
